# api/openapi/order.yml
openapi: 3.0.3
info:
  title: order service
  description: order service
  version: 1.0.0
servers:
    - url: "http://{hostname}/api"
      variables:
        hostname:
          default: 127.0.0.1

paths:
  /customer/{customer_id}/orders/{order_id}:
    get:
      description: "get order"
      parameters:
        - in: path
          name: customer_id
          schema:
            type: string
          required: true

        - in: path
          name: order_id
          schema:
            type: string
          required: true

      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

        default:
          description: "unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /customer/{customer_id}/orders:
    post:
      description: "create order"
      parameters:
        - in: path
          name: customer_id
          schema:
            type: string
          required: true

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'

      responses:
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

        default:
          description: "unexpected error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    OrderStatus:
      type: string
      enum:
        - ORDER_STATUS_UNSPECIFIED
        - ORDER_STATUS_PENDING
        - ORDER_STATUS_PAID
        - ORDER_STATUS_PREPARING
        - ORDER_STATUS_READY
        - ORDER_STATUS_DELIVERING
        - ORDER_STATUS_DELIVERED
        - ORDER_STATUS_CANCELLED
      description: "订单状态"
    
    Order:
      type: object
      required:
        - id
        - customer_id
        - status
        - items
        - payment_link
      properties:
        id:
          type: string
        customer_id:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        payment_link:
          type: string
          description: "支付链接"

    Item:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        quantity:
          type: integer
          format: int32
        price_id:
          type: string

    Error:
        type: object
        required:
            - message
        properties:
          message:
            type: string

    CreateOrderRequest:
      type: object
      required:
        - customer_id
        - items
      properties:
        customer_id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemWithQuantity'

    ItemWithQuantity:
        type: object
        required:
            - id
            - quantity
        properties:
          id:
            type: string
          quantity:
            type: integer
            format: int32


    Response:
        type: object
        required:
            - errno
            - message
            - data
            - trace_id
        properties:
            errno:
                type: integer
                format: int32
            message:
                type: string
            data:
                type: object
            trace_id:
                type: string