# internal/order/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package ports

// Defines values for OrderStatus.
const (
	ORDERSTATUSCANCELLED   OrderStatus = "ORDER_STATUS_CANCELLED"
	ORDERSTATUSDELIVERED   OrderStatus = "ORDER_STATUS_DELIVERED"
	ORDERSTATUSDELIVERING  OrderStatus = "ORDER_STATUS_DELIVERING"
	ORDERSTATUSPAID        OrderStatus = "ORDER_STATUS_PAID"
	ORDERSTATUSPENDING     OrderStatus = "ORDER_STATUS_PENDING"
	ORDERSTATUSPREPARING   OrderStatus = "ORDER_STATUS_PREPARING"
	ORDERSTATUSREADY       OrderStatus = "ORDER_STATUS_READY"
	ORDERSTATUSUNSPECIFIED OrderStatus = "ORDER_STATUS_UNSPECIFIED"
)

// Valid indicates whether the value is a known member of the OrderStatus enum.
func (e OrderStatus) Valid() bool {
	switch e {
	case ORDERSTATUSCANCELLED:
		return true
	case ORDERSTATUSDELIVERED:
		return true
	case ORDERSTATUSDELIVERING:
		return true
	case ORDERSTATUSPAID:
		return true
	case ORDERSTATUSPENDING:
		return true
	case ORDERSTATUSPREPARING:
		return true
	case ORDERSTATUSREADY:
		return true
	case ORDERSTATUSUNSPECIFIED:
		return true
	default:
		return false
	}
}

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	CustomerId string             `json:"customer_id"`
	Items      []ItemWithQuantity `json:"items"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// Item defines model for Item.
type Item struct {
	Id       *string `json:"id,omitempty"`
	Name     *string `json:"name,omitempty"`
	PriceId  *string `json:"price_id,omitempty"`
	Quantity *int32  `json:"quantity,omitempty"`
}

// ItemWithQuantity defines model for ItemWithQuantity.
type ItemWithQuantity struct {
	Id       string `json:"id"`
	Quantity int32  `json:"quantity"`
}

// Order defines model for Order.
type Order struct {
	CustomerId string `json:"customer_id"`
	Id         string `json:"id"`
	Items      []Item `json:"items"`

	// PaymentLink 支付链接
	PaymentLink string `json:"payment_link"`

	// Status 订单状态
	Status OrderStatus `json:"status"`
}

// OrderStatus 订单状态
type OrderStatus string

// Response defines model for Response.
type Response struct {
	Data    map[string]interface{} `json:"data"`
	Errno   int32                  `json:"errno"`
	Message string                 `json:"message"`
	TraceId string                 `json:"trace_id"`
}

// PostCustomerCustomerIdOrdersJSONRequestBody defines body for PostCustomerCustomerIdOrders for application/json ContentType.
type PostCustomerCustomerIdOrdersJSONRequestBody = CreateOrderRequest